Split terminal gutter host into its own file
